+++ countries_plus/__init__.py
@@ -1,3 +1,3 @@
-APP_NAME = 'languages_plus'
+APP_NAME = 'countries_plus'
 default_app_config = 'countries_plus.apps.DefaultConfig'
 __version__ = '1.2.1'