+++ tests/test_models.py
@@ -1,13 +1,14 @@
 from copy import deepcopy
 from decimal import Decimal
 
+import pytest
 from django.core.exceptions import ValidationError
-from django.test import TestCase
 
 from countries_plus.models import Country
 
 
-class TestCountryModel(TestCase):
+@pytest.mark.django_db
+class TestCountryModel:
     valid_values = {
         'postal_code_regex': 'test', 'geonameid': 1149361,
         'languages': 'fa-AF,ps,uz-AF,tk', 'equivalent_fips_code': '2134',
@@ -18,17 +19,49 @@ class TestCountryModel(TestCase):
         'tld': '.af', 'iso': 'AF', 'phone': '9342342 and 4293432434', 'currency_code': 'AFN'
     }
 
-    def setUp(self):
-        Country.objects.all().delete()
-
-    def tearDown(self):
-        Country.objects.all().delete()
-
     def test_country_validation_valid(self):
-        self.assertIsInstance(Country.objects.create(**self.valid_values), Country)
+        assert isinstance(Country.objects.create(**self.valid_values), Country)
 
     def test_country_validation_invalid(self):
         invalid_values = deepcopy(self.valid_values)
         invalid_values['iso'] = "This is too long for the iso field"
-        with self.assertRaises(ValidationError):
+        with pytest.raises(ValidationError):
             Country.objects.create(**invalid_values)
+
+
+@pytest.mark.django_db
+class TestCountryGetByRequest:
+
+    @pytest.fixture
+    def request_without_geoip(self, rf):
+        request = rf.get('/')
+        request.META = {}
+        return request
+
+    @pytest.fixture
+    def request_with_geoip(self, rf):
+        request = rf.get('/')
+        request.META = {
+            'GEOIP_HEADER': 'TC',
+        }
+        return request
+
+    def test_blank_settings(self, settings, request_without_geoip):
+        # Test with missing/badly formed settings
+        settings.COUNTRIES_PLUS_COUNTRY_HEADER = ''
+        settings.COUNTRIES_PLUS_DEFAULT_ISO = ''
+        with pytest.raises(AttributeError):
+            Country.get_by_request(request_without_geoip)
+
+    def test_blank_default(self, settings, request_without_geoip):
+        # Test without a default
+        settings.COUNTRIES_PLUS_COUNTRY_HEADER = 'GEOIP_HEADER'
+        settings.COUNTRIES_PLUS_DEFAULT_ISO = ''
+        assert Country.get_by_request(request_without_geoip) is None
+
+    def test_default(self, settings, default_country, other_country, request_without_geoip, request_with_geoip):
+        # Test with a default
+        settings.COUNTRIES_PLUS_COUNTRY_HEADER = 'GEOIP_HEADER'
+        settings.COUNTRIES_PLUS_DEFAULT_ISO = 'US'
+        assert Country.get_by_request(request_without_geoip) == default_country
+        assert Country.get_by_request(request_with_geoip) == other_country